requirements.txt
gunicorn==20.0.4
jsonschema==3.2.0
python-dateutil==2.8.1
requests==2.13.0
Scrapy==2.0.0
Twisted==19.10.0
typing==3.7.4.1
w3lib==1.21.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.40.0
Werkzeug==0.16.1
Flask==1.1.1
postgres==3.0.0
SQLAlchemy==1.3.13
psycopg2-binary==2.8.4
psycopg2-pool==1.1